Moved fareType switch statement into apiUtils and moved relevant tests
