+++ src/pages/api/fareType.ts
@@ -1,5 +1,12 @@
 import { NextApiRequest, NextApiResponse } from 'next';
-import { getDomain, setCookieOnResponseObject, redirectToError, redirectTo, getUuidFromCookie } from './apiUtils/index';
+import {
+    getDomain,
+    setCookieOnResponseObject,
+    redirectToError,
+    redirectTo,
+    redirectOnFareType,
+    getUuidFromCookie,
+} from './apiUtils/index';
 import { FARETYPE_COOKIE } from '../../constants/index';
 
 import { isSessionValid } from './service/validator';
@@ -19,22 +26,7 @@ export default (req: NextApiRequest, res: NextApiResponse): void => {
 
             setCookieOnResponseObject(getDomain(req), FARETYPE_COOKIE, cookieValue, req, res);
 
-            switch (req.body.fareType) {
-                case 'period':
-                    redirectTo(res, '/periodType');
-                    return;
-                case 'single':
-                    redirectTo(res, '/service');
-                    return;
-                case 'return':
-                    redirectTo(res, '/service');
-                    return;
-                case 'flatFare':
-                    redirectTo(res, '/serviceList');
-                    return;
-                default:
-                    throw new Error('Fare Type we expect was not received.');
-            }
+            redirectOnFareType(req, res);
         } else {
             const cookieValue = JSON.stringify({
                 errorMessage: 'Choose a fare type from the options',
@@ -44,7 +36,7 @@ export default (req: NextApiRequest, res: NextApiResponse): void => {
             redirectTo(res, '/fareType');
         }
     } catch (error) {
-        const message = 'There was a problem selecting the fare type:';
+        const message = 'There was a problem selecting the fare type.';
         redirectToError(res, message, error);
     }
 };
